client/requirements.txt
pyzmq==25.1.1				# PyZMQ for ZeroMQ bindings in Python
Click==8.1.3				# Click for creating command-line interfaces
Sphinx==5.3.0				# Sphinx for generating documentation
python-dotenv==1.0.1